requirements.txt
streamlit
google-cloud-firestore
vega-datasets
pydeck
